requirements.txt
Django>=5.1.3
psycopg2-binary
Pillow
python-telegram-bot==20.8
gunicorn
django-crispy-forms
django-widget-tweaks